Изменил на float дистанцию и время.

+++ Example008_2broAndDog/Program.cs
@@ -1,12 +1,12 @@
 ﻿// Два друга идут навстречу друг другу - дистанция 10000, скорость: друг1 = 1, друг2 = 2, собака = 5. 
 // Сколько раз собака пробежит между друзьями?
 int count = 0;
-double distance = 10000;
+float distance = 10000;
 int firstFriendSpeed = 1;
 int secondFriendSpeed = 2;
 int dogSpeed = 5;
 int friend = 2;
-double time = 0;
+float time = 0;
 while (distance > 10)
 {
 if (friend == 1)
